got rid of timer()

+++ bbs/sr.h
@@ -31,7 +31,7 @@ enum xfertype {
 };
 
 void calc_CRC(unsigned char b);
-char gettimeout(double d, bool *abort);
+char gettimeout(long d, bool *abort);
 int  extern_prot(int nProtocolNum, const char *pfile_nameToSend, bool bSending);
 bool ok_prot(int nProtocolNum, xfertype xt);
 char *prot_name(int nProtocolNum);
